janky fix to !add/say meme
multi line entries broke the command because it split entries in the .txt by linebreaks. this makes it add ~END to the end of memes, and splits it there instead.
obviously can still be broken by typing ~END followed by a line break, but at least multiple lines will work now.

unfortunately this means that any current entries to meme.txt will have to have " ~END" added on for it to work.